==> notebooks/2. ground-truth-data.ipynb <==
# Import Libraries

import os
import json
import hashlib
from collections import defaultdict
import openai
from openai import OpenAI
from tqdm.auto import tqdm
import pandas as pd

# Load JSON file

documents = []
with open('../data/data.json', 'r') as file:
    for line in file:
        documents.append(json.loads(line))

documents[5]

def generate_document_id(doc):
    combined = f"{doc['question']}-{doc['answer']}-{doc['source']}-{doc['focus_area'][:10]}"
    hash_object = hashlib.md5(combined.encode())
    hash_hex = hash_object.hexdigest()
    document_id = hash_hex[:8]
    return document_id

# Generate IDs for each document

for doc in documents:
    doc['id'] = generate_document_id(doc)

documents[5]

# Create a defaultdict to store documents by ID
hashes = defaultdict(list)

# Group documents by their ID
for index,doc in enumerate(documents):
    doc_id = doc['id']
    hashes[doc_id].append(doc)

len(hashes), len(documents)

with open('../data/data-with-ids.json', 'wt') as f_out:
    json.dump(documents, f_out, indent=2)

# Generate the prompt 

prompt_template = """
You emulate a user of our health assistant application.
Formulate 5 questions this user might ask based on a disease.
Make the questions specific to this disease.
The record should contain the answer to the questions, and the questions should
be complete and not too short. Use as fewer words as possible from the record. 

The record:

question: {question}
answer: {answer}
source: {source}
focus_area: {focus_area}

Provide the output in parsable JSON without using code blocks:

{{"questions": ["question1", "question2", ..., "question5"]}}
""".strip()

prompt = prompt_template.format(**documents[5])

api_key = os.getenv("OPENAI_API_KEY")

# Initializing OPENAI client
client = OpenAI()

def llm(prompt):
    response = client.chat.completions.create(
        model='gpt-4o-mini',
        messages=[{"role": "user", "content": prompt}]
    )
    
    return response.choices[0].message.content

questions = llm(prompt)
json.loads(questions)

# Function that sends a request to the OpenAI API to generate questions based on the formatted prompt.

def generate_questions(doc):
    prompt = prompt_template.format(**doc)

    response = client.chat.completions.create(
        model='gpt-4o-mini',
        messages=[{"role": "user", "content": prompt}]
    )

    json_response = response.choices[0].message.content
    return json_response

results = {}

results = {}
for doc in tqdm(documents): 
    doc_id = doc['id']
    if doc_id in results:
        continue

    questions_raw = generate_questions(doc)
    questions = json.loads(questions_raw)
    results[doc_id] = questions['questions']

final_results = []

for doc_id, questions in results.items():
    for q in questions:
        final_results.append((doc_id, q))

final_results[5]

# Save the results to a csv 

df_results = pd.DataFrame(final_results, columns=['id', 'question'])
df_results.to_csv('../data/ground-truth-retrieval.csv', index=False)

!tail ../data/ground-truth-retrieval.csv
