==> notebooks/3. text-search-eval.ipynb <==
# Import libraries

import json
import pandas as pd
from elasticsearch import Elasticsearch, exceptions
from tqdm.auto import tqdm
import minsearch

import warnings
warnings.filterwarnings("ignore")

# Read JSON document with generated ids

with open('..\data\data-with-ids.json', 'rt') as f_in:
    documents = json.load(f_in)

# Load generated ground truth data

df_questions = pd.read_csv('..\data\ground-truth-retrieval.csv')
ground_truth = df_questions.to_dict(orient='records')

ground_truth[100]

#### **Elastic Search**

es = Elasticsearch('http://localhost:9200') 

# Define index name and settings
index_name = "health-questions"

index_settings = {
    "settings": {
        "number_of_shards": 1,
        "number_of_replicas": 0
    },
    "mappings": {
        "properties": {
            "question": {"type": "text"},
            "answer": {"type": "text"},
            "source": {"type": "text"},
            "focus_area": {"type": "text"}
        }
    }
}

# Create the index if it doesn't exist

try:
    if not es.indices.exists(index=index_name):
        es.indices.create(index=index_name, body=index_settings)
        print("Index created successfully")
    else:
        print("Index already exists")
except exceptions.ConnectionError as e:
    print(f"Failed to create index: {e}")

# Index the documents
for doc in tqdm(documents):
    es.index(index=index_name, document=doc)

# Define a function that retrieves documents and matches user queries

def elastic_search1(query, max_results=5):
    search_query = {
        "size": max_results,
        "query": {
            "bool": {
                "must": {
                    "multi_match": {
                        "query": query,
                        "fields": ["question", "answer", "focus_area"],   # no boosting
                        "type": "best_fields"
                    }
                }
            }
        }
    }
    
    response = es.search(index=index_name, body=search_query)
    documents = [hit['_source'] for hit in response['hits']['hits']]
    return documents

# Define a function that retrieves documents and matches user queries with boosting

def elastic_search2(query, max_results=5):
    search_query = {
        "size": max_results,
        "query": {
            "bool": {
                "must": {
                    "multi_match": {
                        "query": query,
                        "fields": ["question^3", "answer^3", "focus_area^2"],  # boosting
                        "type": "best_fields"
                    }
                }
            }
        }
    }
    
    response = es.search(index=index_name, body=search_query)
    documents = [hit['_source'] for hit in response['hits']['hits']]
    return documents

#### **Minsearch**

index = minsearch.Index(
    text_fields=["question", "answer", "focus_area", "source"],
    keyword_fields=["id"]
)

index.fit(documents)

def minsearch_search1(query):
    boost = {} # no boosting

    results = index.search(
        query=query,
        boost_dict=boost,
        num_results=5
    )

    return results

def minsearch_search2(query):
    boost = {'question': 3.0, 'answer' : 3.0, 'focus_area' : 2.0} # with boosting

    results = index.search(
        query=query,
        boost_dict=boost,
        num_results=5
    )

    return results

#### Evaluation metrics :
  - *Hit Rate (recall)*
  - *Mean Reciprocal Rank (mrr)*

def hit_rate(relevance_total):
    cnt = 0

    for line in relevance_total:
        if True in line:
            cnt = cnt + 1

    return cnt / len(relevance_total)

def mrr(relevance_total):
    total_score = 0.0

    for line in relevance_total:
        for rank in range(len(line)):
            if line[rank]:  # If relevant document found
                total_score += 1 / (rank + 1)
                break  # Stop after finding the first relevant document

    return total_score / len(relevance_total)

#### Retrieval Evaluation

def evaluate(ground_truth, search_function):
    relevance_total = []

    for q in tqdm(ground_truth):
        doc_id = q['id']
        results = search_function(q)
        relevance = [d['id'] == doc_id for d in results]
        relevance_total.append(relevance)

    return {
        'hit_rate': hit_rate(relevance_total),
        'mrr': mrr(relevance_total),
    }

# Elastic search
evaluate(ground_truth, lambda q: elastic_search1(q['question']))

evaluate(ground_truth, lambda q: elastic_search2(q['question']))

# Minsearch
evaluate(ground_truth, lambda q: minsearch_search1(q['question']))

evaluate(ground_truth, lambda q: minsearch_search2(q['question']))

Minsearch produces better results than Elastic Search with and with no boosting
